backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-magic==0.4.27
Pillow==10.1.0
reportlab==4.0.7
pdf2docx==0.5.6
PyMuPDF==1.23.7
docx2pdf==0.1.8
python-docx==0.8.11
python-pptx==0.6.21
pypdf==3.17.1
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
openpyxl==3.1.2
